+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr 0.2.0.9000
 
+* `nth` now correctly promotes the result when using dates and times (#509). 
+
 * more coherence when joining columns of compatible but different types, e.g. when
   joining a character vector and a factor (#455).